\name{DemoPackage}
\alias{DemoPackage}
\docType{package}
\title{DemoPackage:  a minimal package, to adopt and adapt for your own work.
}
\description{
yabba!
}
\details{
\tabular{ll}{
Package: \tab demo\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-09\cr
License: \tab LGPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Paul Shannon <pshannon@systemsbiology.org>
Maintainer: Paul Shannon <pshannon@systemsbiology.org>
}
\references{
none yet...
}
\keyword{dynamic}
\seealso{
  NA
}
\examples{
  NA
}
